Alias "Invalid story id" to a `StoryError` variant
The only one that I have seen is this one, so I add it as variant to
make it easier to handle it.

While the `StoryError::Api(_)` is kept as a catch-all, the "error" key
might not always contain a string. Finding another error from API is
needed to test this.