Remove `Unpublished` from `StoryError`
The `from_str()` function shouldn't care if the content is "incomplete"
when it successfully deserialized the content. As shown by the fact that
it wasn't being discarded and instead returned as an error.

Let the user decide if that actually represents an error state for their
use case by removing that error variant and providing the unpublished
check as a method. Getting the bonus of reducing the size of the error
itself.